Rename "Hidden" Movement Type to "Buried"
This commit mirrors the change to pokeemerald that renames `MOVEMENT_TYPE_HIDDEN` to `MOVEMENT_TYPE_BURIED` to avoid confusion between it and `MOVEMENT_TYPE_INVISIBLE`.

As the files and function labels are not synced between the two repos, only the necessary changes were made based on what pokefirered calls its functions. More syncing of labels with pokeemerald will need to be done in the future.