Fix Typos / Use Box Constants
The Celadon City scripts that give Eevee misspelled it as Eeevee. Fixed.

The comment explaining `boxRSEggsUnlocked` in the `ExternalEventFlags` struct had an incorrect deposit requirement listed and was fixed.

Also, replaced the literals with the `TOTAL_BOXES_COUNT` and `IN_BOX_COUNT` constants in `pokemon.c`'s `IsPokemonStorageFull` function. Thanks hjk321 for pointing that out.